backend/requirements.txt
uvicorn
python-multipart
fastapi
matplotlib
Pillow
